+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ### HEAD
 
+### 0.5.2 (August 19, 2017)
+
+  * Completed the library with new patterns functional, prototypal and pseudoclassical,
+  * Set functional-shared as the default pattern,
+  * ...,
+
+
 ### 0.5.1 (August 17, 2017)
 
   * Forced the library file name to always be in lowercase characters,